CAM: rename Shape module to shape (PEP8)
