+++ src/Mod/CAM/Path/Tool/shape/__init__.py
