+++ src/Mod/CAM/Path/Tool/shape/base.py
