+++ src/Mod/CAM/Path/Tool/shape/chamfer.py
