+++ src/Mod/CAM/Path/Tool/shape/doc.py
