+++ src/Mod/CAM/Path/Tool/shape/endmill.py
