+++ src/Mod/CAM/Path/Tool/shape/slittingsaw.py
