+++ src/Mod/CAM/Path/Tool/shape/util.py
