+++ src/Mod/CAM/Path/Tool/shape/vbit.py
